fix requested changes
